requirements.txt
# Core dependencies for SlideAudioGenerator
google-generativeai>=0.3.0
pyinstaller>=5.0
